requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-genai>=1.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiosmtplib>=3.0.0
Pillow>=10.0.0
img2pdf>=0.5.0
pydantic-settings>=2.0.0
cryptography>=41.0.0
